refactor(UserChallenges): Renaming and moving User/Challenges to UserChallenges
