+++ src/modules/App/App.middleware.ts
@@ -2,8 +2,8 @@ import { Middleware } from 'redux'
 
 // avoiding circular dependencies:
 import * as AuthActions from '../Auth/Auth.reducer'
-import * as UserChallengesActions from '../User/Challenges/Challenges.reducer'
 import * as UserActions from '../User/User.reducer'
+import * as UserChallengesActions from '../UserChallenges/UserChallenges.reducer'
 import { hydrateApp, resetAppData } from './App.reducer'
 
 export const appResetFlow: Middleware =
