--- src/modules/User/Challenges/index.ts
@@ -1,5 +1,0 @@
-import * as middleware from './Challenges.middleware'
-import reducer, * as actions from './Challenges.reducer'
-import * as types from './Challenges.types'
-
-export { actions, reducer, middleware, types }
