+++ src/modules/UserChallenges/index.ts
@@ -1,0 +1,5 @@
+import * as middleware from './UserChallenges.middleware'
+import reducer, * as actions from './UserChallenges.reducer'
+import * as types from './UserChallenges.types'
+
+export { actions, reducer, middleware, types }
