+++ src/redux/reducers.ts
@@ -2,7 +2,7 @@ import { combineReducers } from 'redux'
 
 import { reducer as auth } from '../modules/Auth'
 import { reducer as user } from '../modules/User'
-import { reducer as userChallenges } from '../modules/User/Challenges'
+import { reducer as userChallenges } from '../modules/UserChallenges'
 
 const rootReducer = combineReducers({
   auth,
