+++ src/redux/redux.test.fixtures.ts
@@ -1,7 +1,7 @@
 import { createFixture } from '../../tests/tests.utils'
 import { INITIAL_STATE as AUTH_INITIAL_STATE } from '../modules/Auth/Auth.reducer'
-import { INITIAL_STATE as USER_CHALLENGES_INITIAL_STATE } from '../modules/User/Challenges/Challenges.reducer'
 import { INITIAL_STATE as USER_INITIAL_STATE } from '../modules/User/User.reducer'
+import { INITIAL_STATE as USER_CHALLENGES_INITIAL_STATE } from '../modules/UserChallenges/UserChallenges.reducer'
 import { RootState } from './redux.types'
 
 export const defaultRootState: RootState = {
