+++ src/utils/redux.utils.ts
@@ -1,4 +1,3 @@
-import { PayloadAction } from '@reduxjs/toolkit'
 import { applySpec, identity, keys, path, pipe, prop } from 'ramda'
 
 import { objFromListWith } from './ramda.utils'
